Restructure to umbrella app
