--- README.md
@@ -1,21 +1,0 @@
-# Wallet App
-
-Manage mutli currency wallets
-
-## Contexts
-
-Accounts
-
-  - [x] Register
-  - [x] Login
-
-Wallets
-
-  - [x] Create Wallet
-  - [x] Get user wallets
-  - [x] Get user wallet
-  - [x] Get user wallet transactions
-
-Transaction
-
-  - [ ] Create wallet transaction
