+++ apps/wallet/lib/wallet_app/accounts/user.ex
