+++ apps/wallet/lib/wallet_app/application.ex
