+++ apps/wallet/lib/wallet_app/wallets/wallet.ex
