+++ config/config.exs
@@ -2,12 +2,16 @@
 # and its dependencies with the aid of the Mix.Config module.
 use Mix.Config
 
-config :wallet_app, WalletApp.Repo,
-  adapter: Sqlite.Ecto2,
-  database: "wallet.sqlite3"
+# By default, the umbrella project as well as each child
+# application will require this configuration file, ensuring
+# they all use the same configuration. While one could
+# configure all applications here, we prefer to delegate
+# back to each application for organization purposes.
+import_config "../apps/*/config/config.exs"
 
-config :wallet_app, ecto_repos: [WalletApp.Repo]
-
-config :wallet_app, jwt_alg: "HS256", jwt_key: "Bs0PzP3VV5pHtaE4M4nJblvnVphq6oVS"
-
-import_config "#{Mix.env()}.exs"
+# Sample configuration (overrides the imported configuration above):
+#
+#     config :logger, :console,
+#       level: :info,
+#       format: "$date $time [$level] $metadata$message\n",
+#       metadata: [:user_id]
